requirements.txt
flask
click
google-cloud-dns
google-cloud-datastore
